+++ CHANGELOG.md
@@ -1,11 +1,25 @@
 ## Unreleased
 
+## 5.0.0
+
 ### Enhancements
 * Set multiple attribute for grouped selects also. [@ollym](https://github.com/ollym)
 * Removes or renames label classes. [Abduvakilov](https://github.com/Abduvakilov)
 * Support to label custom classes for inline collections. [@feliperenan](https://github.com/feliperenan)
 * Update bootstrap generator template to match v4.3.x. [@m5o](https://github.com/m5o)
 * Allow "required" attribute in generated select elements of PriorityInput. [@mcountis](https://github.com/mcountis)
+
+### Bug fix
+* Do not call `#send` in form object to check whether the attribute is a file input. [@tegon](https://github.com/tegon)
+
+## Deprecations
+* The config `SimpleForm.file_methods` is deprecated and it has no effect. Simple Form now supports automatically discover of file inputs for the following Gems: activestorage, carrierwave, paperclip, refile and shrine. If you are using a custom method that is not from one of the supported Gems, please change your forms to pass the input type explicitly:
+
+```erb
+ <%= form.input :avatar, as: :file %>
+ ```
+
+See http://blog.plataformatec.com.br/2019/09/incorrect-access-control-in-simple-form-cve-2019-16676 for more information.
 
 ## 4.1.0